Revise the exit codes returned by the program
Exit code 2 means the program failed to perform the requested operation.
It is the most important exit code so it should be number 1.

Exit code 1 means the program failed to open the USB device.
This usually means the program didn't have the required permissions.
Elevated privileges are usually required to access the USB devices.
Although it is probably the most common exit code, it is not related
to the operation requested by the user. So it should be number 2.