Refactor brightness and speed setting functions
Turns out the unknown byte that follows the brightness is the speed.
Both values are sent in the same feature report.
The brightness and speed functions have been refactored to reflect this.
A new function that sets both at the same time has also been introduced.

This design makes it more complicated to set them separately.
Care must be taken not to override the previously set value.
To handle this, the program now keeps track of the brightness and speed.
When the program sets one, it also sets the other to its current value.
The result is only one of the parameters is changed changed.

There seems to be no way to obtain the state of the LEDs
from the keyboard itself. So when the program starts
the brightness and speed are unknown and assumed to be zero.
The user is advised to set all these parameters explicitly.