Check for NULL and empty inputs
The caller may pass null or empty commands lists, command lines
and argument vectors. Since there is nothing to process,
the functions should return successfully without doing anything.

Unlike empty inputs, the null inputs may or may not be an error.
It is up to the user to decide by checking for NULL before calling.